+++ packages/next-common/components/charts/thresholdCurve/legend/gov2TallyLegend.js
